Prevent division by zero
In running a small test sample I encountered a failure when the script attempted a division by zero.

Perhaps my sample was so small that some of the things  the script is counting had zero occurrences.

The update sets the values of certain "average" variables to zero if the set being averaged has no members. I hope that is in keeping with the calculations you are performing.